ambd_flash_tool/requirements.txt
Click
colorama
pyserial